+++ kem/schemes/schemes.go
@@ -1,3 +1,5 @@
+//go:build !wasm
+
 package schemes
 
 import (
