+++ nike/hybrid/ctidh.go
@@ -1,6 +1,8 @@
 // SPDX-FileCopyrightText: (c) 2024 David Stainton
 // SPDX-License-Identifier: AGPL-3.0-only
 
+//go:build !wasm
+
 package hybrid
 
 import (
