+++ sign/sphincsplus/sphincs.go
@@ -1,4 +1,4 @@
-//go:build !windows
+//go:build !windows && !wasm
 
 // SPDX-FileCopyrightText: (c) 2022-2024 David Stainton
 // SPDX-License-Identifier: AGPL-3.0-only
